* changed TSP distances from Integer to Double + cleanUp method for algorithms * fixed bug in GE * minor changes
git-svn-id: svn://svn.code.sf.net/p/jcop/code/trunk@33 ca20af45-4a20-473b-94bf-0ef87cbc3059

+++ jcop/tests/cz/cvut/felk/cig/jcop/problem/tsp/TSPIteratorTest.java
@@ -17,10 +17,10 @@ import org.testng.annotations.Test;
 public class TSPIteratorTest {
     @Test
     public void testNext() throws Exception {
-        TSP problem = new TSP(new Integer[][]{
-                {0, 1, 2},
-                {3, 0, 4},
-                {5, 6, 0},
+        TSP problem = new TSP(new Double[][]{
+                {0.0, 1.0, 2.0},
+                {3.0, 0.0, 4.0},
+                {5.0, 6.0, 0.0},
         });
 
         Configuration configuration = problem.getStartingConfiguration();
